Change ht_key_t typedef to size_t
This allows a larger hash table. The bit width of keys no longer limited
to 2^32 in 64-bit builds.

+++ Action.c
@@ -113,7 +113,7 @@ static bool changePriority(MainPanel* panel, int delta) {
 static void addUserToVector(ht_key_t key, void* userCast, void* panelCast) {
    const char* user = userCast;
    Panel* panel = panelCast;
-   Panel_add(panel, (Object*) ListItem_new(user, key));
+   Panel_add(panel, (Object*) ListItem_new(user, (int)key));
 }
 
 bool Action_setUserOnly(const char* userName, uid_t* userId) {
